Add files via upload
In this project, the student group (3-4 students) will spend time creating an e-commerce website. Choose a product of your choice.
The project should have a basic e-Commerce website. A user should be able to navigate through a number of products. The user should be able to use filters on products.
The user should be able to add products to a cart. Before checkout, the user should be able to register. At checkout, the user should get a pdf invoice. (implementing a payment gateway is the future scope).
Create a SQL database according to the requirements of the website. The database should be normalized to at least the 3rd Normal Form.
Use all the topics covered in this course to complete this project.
Your website should implement but not limited to:
•	CRUD operation using Object-Oriented Programming in PHP. If Object-Oriented Programming is not used, then 50% marks will be deducted.
•	Use effective techniques for preventing SQL Injection and XSS and other potential threats.
•	Effective use of Cookies and Sessions.
•	Use of APIs.
•	Should have at least 4 pages.
•	Use appropriate colors and design elements.